flask/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
ultralytics==8.2.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4
torch==2.1.2
torchvision==0.16.2
gunicorn==21.2.0
